+++ packages/backend/src/core/assets/index.ts
@@ -1,0 +1,2 @@
+export * from './CBVUpdater'
+export * from './NMVUpdater'
